Refactor category and subcategory controllers and services to move hits subcategory to subcategory module
